requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Mail==0.9.1
stripe==5.12.0
pyserial==3.5
requests==2.31.0
Flask-APScheduler==1.12.3